+++ Common/src/main/java/fuzs/visualworkbench/config/JsonConfigBuilder.java
@@ -1,24 +1,24 @@
 package fuzs.visualworkbench.config;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Sets;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
-import fuzs.puzzleslib.core.CoreServices;
 import fuzs.puzzleslib.json.JsonConfigFileUtil;
 import fuzs.puzzleslib.json.JsonSerializationUtil;
 import fuzs.visualworkbench.VisualWorkbench;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.CraftingTableBlock;
 
 import java.io.File;
 import java.io.FileReader;
 import java.lang.reflect.Type;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -30,7 +30,9 @@ public class JsonConfigBuilder {
             .create();
     private static final String JSON_CONFIG_NAME = "visual_workbench.json";
     private static final List<ResourceLocation> DEFAULT_VISUAL_WORKBENCHES = Stream.of(
+                    // Minecraft
                     "minecraft:crafting_table",
+                    // Blue Skies
                     "blue_skies:bluebright_crafting_table",
                     "blue_skies:starlit_crafting_table",
                     "blue_skies:frostbright_crafting_table",
@@ -38,6 +40,7 @@ public class JsonConfigBuilder {
                     "blue_skies:dusk_crafting_table",
                     "blue_skies:maple_crafting_table",
                     "blue_skies:cherry_crafting_table",
+                    // Blocks +
                     "blocksplus:spruce_crafting_table",
                     "blocksplus:birch_crafting_table",
                     "blocksplus:jungle_crafting_table",
@@ -47,6 +50,7 @@ public class JsonConfigBuilder {
                     "blocksplus:warped_crafting_table",
                     "blocksplus:bamboo_crafting_table",
                     "blocksplus:mushroom_crafting_table",
+                    // More Crafting Tables for Forge!
                     "mctb:spruce_crafting_table",
                     "mctb:birch_crafting_table",
                     "mctb:acacia_crafting_table",
@@ -65,27 +69,43 @@ public class JsonConfigBuilder {
                     "mctb:redwood_crafting_table",
                     "mctb:umbran_crafting_table",
                     "mctb:willow_crafting_table",
-                    "betternether:crafting_table_anchor_tree",
-                    "betternether:crafting_table_nether_sakura",
+                    "mctb:azalea_crafting_table",
+                    "mctb:blossom_crafting_table",
+                    // BetterNether
+                    "betternether:rubeus_crafting_table",
+                    "betternether:nether_sakura_crafting_table",
                     "betternether:crafting_table_crimson",
+                    "betternether:wart_crafting_table",
                     "betternether:crafting_table_warped",
-                    "betternether:crafting_table_stalagnate",
-                    "betternether:crafting_table_reed",
-                    "betternether:crafting_table_willow",
-                    "betternether:crafting_table_wart",
-                    "betternether:crafting_table_rubeus",
-                    "betternether:crafting_table_mushroom",
-                    "betternether:crafting_table_mushroom_fir",
+                    "betternether:anchor_tree_crafting_table",
+                    "betternether:willow_crafting_table",
+                    "betternether:nether_mushroom_crafting_table",
+                    "betternether:stalagnate_crafting_table",
+                    "betternether:mushroom_fir_crafting_table",
+                    "betternether:nether_reed_crafting_table",
+                    // BetterEnd
+                    "betterend:helix_tree_crafting_table",
+                    "betterend:mossy_glowshroom_crafting_table",
+                    "betterend:end_lotus_crafting_table",
+                    "betterend:pythadendron_crafting_table",
+                    "betterend:jellyshroom_crafting_table",
+                    "betterend:tenanea_crafting_table",
+                    "betterend:dragon_tree_crafting_table",
+                    "betterend:lucernia_crafting_table",
+                    "betterend:lacugrove_crafting_table",
+                    "betterend:umbrella_tree_crafting_table",
+                    // BetterEnd Reforked
+                    "betterendforge:helix_tree_crafting_table",
                     "betterendforge:mossy_glowshroom_crafting_table",
-                    "betterendforge:lacugrove_crafting_table",
                     "betterendforge:end_lotus_crafting_table",
                     "betterendforge:pythadendron_crafting_table",
-                    "betterendforge:dragon_tree_crafting_table",
-                    "betterendforge:tenanea_crafting_table",
-                    "betterendforge:helix_tree_crafting_table",
-                    "betterendforge:umbrella_tree_crafting_table",
                     "betterendforge:jellyshroom_crafting_table",
+                    "betterendforge:tenanea_crafting_table",
+                    "betterendforge:dragon_tree_crafting_table",
                     "betterendforge:lucernia_crafting_table",
+                    "betterendforge:lacugrove_crafting_table",
+                    "betterendforge:umbrella_tree_crafting_table",
+                    // Crumbs
                     "crumbs:spruce_crafting_table",
                     "crumbs:birch_crafting_table",
                     "crumbs:jungle_crafting_table",
@@ -93,6 +113,7 @@ public class JsonConfigBuilder {
                     "crumbs:dark_oak_crafting_table",
                     "crumbs:crimson_crafting_table",
                     "crumbs:warped_crafting_table",
+                    // Oh The Biomes You'll Go
                     "byg:aspen_crafting_table",
                     "byg:baobab_crafting_table",
                     "byg:blue_enchanted_crafting_table",
@@ -122,6 +143,7 @@ public class JsonConfigBuilder {
                     "byg:nightshade_crafting_table",
                     "byg:ether_crafting_table",
                     "byg:imparius_crafting_table",
+                    // Variant Crafting Tables [Forge]
                     "vct:spruce_crafting_table",
                     "vct:birch_crafting_table",
                     "vct:jungle_crafting_table",
@@ -189,19 +211,98 @@ public class JsonConfigBuilder {
                     "vct:quark_blossom_crafting_table",
                     "vct:grongle_crafting_table",
                     "vct:smogstem_crafting_table",
-                    "vct:wigglewood_crafting_table"
-                    )
+                    "vct:wigglewood_crafting_table",
+                    // Variant Crafting Tables [Fabric]
+                    "variantcraftingtables:acacia_crafting_table",
+                    "variantcraftingtables:birch_crafting_table",
+                    "variantcraftingtables:dark_oak_crafting_table",
+                    "variantcraftingtables:jungle_crafting_table",
+                    "variantcraftingtables:spruce_crafting_table",
+                    "variantcraftingtables:mangrove_crafting_table",
+                    "variantcraftingtables:crimson_crafting_table",
+                    "variantcraftingtables:warped_crafting_table",
+                    "variantcraftingtables:rubber_crafting_table",
+                    "variantcraftingtables:bamboo_crafting_table",
+                    "variantcraftingtables:charred_crafting_table",
+                    "variantcraftingtables:legacy_crafting_table",
+                    "variantcraftingtables:white_oak_crafting_table",
+                    "variantcraftingtables:herringbone_acacia_crafting_table",
+                    "variantcraftingtables:herringbone_birch_crafting_table",
+                    "variantcraftingtables:herringbone_dark_oak_crafting_table",
+                    "variantcraftingtables:herringbone_jungle_crafting_table",
+                    "variantcraftingtables:herringbone_oak_crafting_table",
+                    "variantcraftingtables:herringbone_spruce_crafting_table",
+                    "variantcraftingtables:herringbone_white_oak_crafting_table",
+                    "variantcraftingtables:herringbone_bamboo_crafting_table",
+                    "variantcraftingtables:herringbone_charred_crafting_table",
+                    "variantcraftingtables:herringbone_crimson_crafting_table",
+                    "variantcraftingtables:herringbone_warped_crafting_table",
+                    "variantcraftingtables:cherry_oak_crafting_table",
+                    "variantcraftingtables:dark_amaranth_crafting_table",
+                    "variantcraftingtables:palm_crafting_table",
+                    "variantcraftingtables:cypress_crafting_table",
+                    "variantcraftingtables:dragons_blood_crafting_table",
+                    "variantcraftingtables:elder_crafting_table",
+                    "variantcraftingtables:juniper_crafting_table",
+                    "variantcraftingtables:dreamwood_crafting_table",
+                    "variantcraftingtables:livingwood_crafting_table",
+                    "variantcraftingtables:mossy_dreamwood_crafting_table",
+                    "variantcraftingtables:mossy_livingwood_crafting_table",
+                    "variantcraftingtables:shimmerwood_crafting_table",
+                    "variantcraftingtables:black_crafting_table",
+                    "variantcraftingtables:blue_crafting_table",
+                    "variantcraftingtables:brown_crafting_table",
+                    "variantcraftingtables:cyan_crafting_table",
+                    "variantcraftingtables:gray_crafting_table",
+                    "variantcraftingtables:green_crafting_table",
+                    "variantcraftingtables:light_blue_crafting_table",
+                    "variantcraftingtables:light_gray_crafting_table",
+                    "variantcraftingtables:lime_crafting_table",
+                    "variantcraftingtables:magenta_crafting_table",
+                    "variantcraftingtables:orange_crafting_table",
+                    "variantcraftingtables:pink_crafting_table",
+                    "variantcraftingtables:purple_crafting_table",
+                    "variantcraftingtables:red_crafting_table",
+                    "variantcraftingtables:white_crafting_table",
+                    "variantcraftingtables:yellow_crafting_table",
+                    "variantcraftingtables:ancient_oak_crafting_table",
+                    "variantcraftingtables:blighted_balsa_crafting_table",
+                    "variantcraftingtables:swamp_cypress_crafting_table",
+                    "variantcraftingtables:willow_crafting_table",
+                    "variantcraftingtables:mango_crafting_table",
+                    "variantcraftingtables:wisteria_crafting_table",
+                    "variantcraftingtables:bamboo_crafting_table_ve",
+                    "variantcraftingtables:redwood_crafting_table",
+                    "variantcraftingtables:azalea_crafting_table",
+                    "variantcraftingtables:coconut_crafting_table",
+                    "variantcraftingtables:flowering_azalea_crafting_table",
+                    "variantcraftingtables:walnut_crafting_table",
+                    "variantcraftingtables:stripped_bamboo_crafting_table",
+                    "variantcraftingtables:crystal_crafting_table",
+                    "variantcraftingtables:golden_oak_crafting_table",
+                    "variantcraftingtables:orange_crafting_table_pl",
+                    "variantcraftingtables:skyroot_crafting_table",
+                    "variantcraftingtables:wisteria_crafting_table_pl",
+                    "variantcraftingtables:cinnamon_crafting_table",
+                    "variantcraftingtables:jade_crafting_table",
+                    "variantcraftingtables:moon_crafting_table",
+                    "variantcraftingtables:shadow_crafting_table"
+            )
             .map(ResourceLocation::new)
-            .collect(Collectors.collectingAndThen(Collectors.toList(), ImmutableList::copyOf));
+            .collect(ImmutableList.toImmutableList());
 
-    private ResourceLocation[] locationValues = {new ResourceLocation("minecraft:crafting_table")};
+    private Set<ResourceLocation> locationValues;
     private Set<Block> blockValues;
 
     public Stream<Block> getBlockStream() {
-        return Stream.of(this.locationValues)
-                .filter(Objects::nonNull)
+        return this.locationValues.stream()
                 .filter(Registry.BLOCK::containsKey)
-                .map(Registry.BLOCK::get);
+                .map(Registry.BLOCK::get)
+                .filter(block -> block instanceof CraftingTableBlock);
+    }
+
+    public boolean isValidCraftingTable(ResourceLocation id, Block block) {
+        return block instanceof CraftingTableBlock && this.locationValues.contains(id);
     }
 
     public boolean contains(Block block) {
@@ -211,12 +312,12 @@ public class JsonConfigBuilder {
         return this.blockValues.contains(block);
     }
 
+    public void clearCache() {
+        this.blockValues = null;
+    }
+
     public void load() {
-        // this doesn't work on Fabric as there is no order in which registries are loaded,
-        // so there is no guarantee modded crafting table blocks will be registered before our block entity is
-        // therefore just hardcode to vanilla table
-        if (!CoreServices.ENVIRONMENT.getModLoader().isForge()) return;
-        JsonConfigFileUtil.getAndLoad(JSON_CONFIG_NAME, JsonConfigBuilder::serialize, (FileReader reader) -> this.locationValues = JsonConfigBuilder.deserialize(reader));
+        JsonConfigFileUtil.getAndLoad(JSON_CONFIG_NAME, JsonConfigBuilder::serialize, (FileReader reader) -> this.locationValues = Sets.newHashSet(JsonConfigBuilder.deserialize(reader)));
     }
 
     private static void serialize(File jsonFile) {