+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.1.1-1.19.1] - 2022-08-14
+### Changed
+- Modded crafting tables are now fully supported on Fabric yay!
+- Added default compat for crafting tables from a few more mods
+
 ## [v4.1.0-1.19.1] - 2022-08-01
 - Compiled for Minecraft 1.19.1
 - Updated to Puzzles Lib v4.1.0